add single recorder
