+++ core/src/main/java/org/apache/carbondata/core/carbon/querystatistics/QueryStatisticsRecorder.java
@@ -20,11 +20,12 @@ package org.apache.carbondata.core.carbon.querystatistics;
 
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 import org.apache.carbondata.common.logging.LogService;
 import org.apache.carbondata.common.logging.LogServiceFactory;
+
+import static org.apache.carbondata.core.carbon.querystatistics.SingleQueryStatisticsRecorder.printLine;
 
 /**
  * Class will be used to record and log the query statistics
@@ -74,42 +75,10 @@ public class QueryStatisticsRecorder implements Serializable {
   }
 
   /**
-   * singleton QueryStatisticsRecorder for driver
-   */
-  private HashMap<String, List<QueryStatistic>> queryStatisticsMap;
-
-  private QueryStatisticsRecorder() {
-    queryStatisticsMap = new HashMap<String, List<QueryStatistic>>();
-  }
-
-  private static QueryStatisticsRecorder carbonLoadStatisticsImplInstance =
-      new QueryStatisticsRecorder();
-
-  public static QueryStatisticsRecorder getInstance() {
-    return carbonLoadStatisticsImplInstance;
-  }
-
-  /**
-   * Below method will be used to add the statistics
-   *
-   * @param statistic
-   */
-  public synchronized void recordStatisticsForDriver(QueryStatistic statistic, String queryId) {
-    // refresh query Statistics Map
-    if (queryStatisticsMap.get(queryId) != null) {
-      queryStatisticsMap.get(queryId).add(statistic);
-    } else {
-      List<QueryStatistic> newQueryStatistics = new ArrayList<QueryStatistic>();
-      newQueryStatistics.add(statistic);
-      queryStatisticsMap.put(queryId, newQueryStatistics);
-    }
-  }
-
-  /**
    * Below method will be used to show statistic log as table
    */
-  public void logStatisticsAsTable(boolean isDriver) {
-    String tableInfo = isDriver? collectDriverStatistics(): collectExecutorStatistics();
+  public void logStatisticsAsTableExecutor() {
+    String tableInfo = collectExecutorStatistics();
     if (null != tableInfo) {
       LOGGER.statistic(tableInfo);
     }
@@ -125,26 +94,27 @@ public class QueryStatisticsRecorder implements Serializable {
     String load_dictionary_time = "";
     String result_size = "";
     String total_executor_time = "";
+    String splitChar = " ";
     try {
       for (QueryStatistic statistic : queryStatistics) {
         switch (statistic.getMessage()) {
           case QueryStatisticsConstants.LOAD_BLOCKS_EXECUTOR:
-            load_blocks_time += statistic.getTimeTaken();
+            load_blocks_time += statistic.getTimeTaken() + splitChar;
             break;
           case QueryStatisticsConstants.SCAN_BLOCKS_TIME:
-            scan_blocks_time += statistic.getTimeTaken();
+            scan_blocks_time += statistic.getTimeTaken() + splitChar;
             break;
           case QueryStatisticsConstants.SCAN_BLOCKS_NUM:
-            scan_blocks_num += statistic.getCount();
+            scan_blocks_num += statistic.getCount() + splitChar;
             break;
           case QueryStatisticsConstants.LOAD_DICTIONARY:
-            load_dictionary_time += statistic.getTimeTaken();
+            load_dictionary_time += statistic.getTimeTaken() + splitChar;
             break;
           case QueryStatisticsConstants.RESULT_SIZE:
-            result_size += statistic.getCount();
+            result_size += statistic.getCount() + " ";
             break;
           case QueryStatisticsConstants.EXECUTOR_PART:
-            total_executor_time += statistic.getTimeTaken();
+            total_executor_time += statistic.getTimeTaken() + splitChar;
             break;
           default:
             break;
@@ -183,114 +153,4 @@ public class QueryStatisticsRecorder implements Serializable {
     }
   }
 
-  /**
-   * Below method will parse queryStatisticsMap and put time into table
-   */
-  public String collectDriverStatistics() {
-    for (String key: queryStatisticsMap.keySet()) {
-      try {
-        // TODO: get the finished query, and print Statistics
-        if (queryStatisticsMap.get(key).size() > 1) {
-          String sql_parse_time = "";
-          String load_meta_time = "";
-          String block_allocation_time = "";
-          String block_identification_time = "";
-          String driver_part_time = "";
-          Double driver_part_time_tmp = 0.0;
-          // get statistic time from the QueryStatistic
-          for (QueryStatistic statistic : queryStatisticsMap.get(key)) {
-            switch (statistic.getMessage()) {
-              case QueryStatisticsConstants.SQL_PARSE:
-                sql_parse_time += statistic.getTimeTaken();
-                driver_part_time_tmp += statistic.getTimeTaken();
-                break;
-              case QueryStatisticsConstants.LOAD_META:
-                load_meta_time += statistic.getTimeTaken();
-                driver_part_time_tmp += statistic.getTimeTaken();
-                break;
-              case QueryStatisticsConstants.BLOCK_ALLOCATION:
-                block_allocation_time += statistic.getTimeTaken();
-                driver_part_time_tmp += statistic.getTimeTaken();
-                break;
-              case QueryStatisticsConstants.BLOCK_IDENTIFICATION:
-                block_identification_time += statistic.getTimeTaken();
-                driver_part_time_tmp += statistic.getTimeTaken();
-                break;
-              default:
-                break;
-            }
-          }
-          driver_part_time = driver_part_time_tmp + "";
-          // structure the query statistics info table
-          StringBuilder tableInfo = new StringBuilder();
-          int len1 = 8;
-          int len2 = 20;
-          int len3 = 21;
-          int len4 = 22;
-          String line = "+" + printLine("-", len1) + "+" + printLine("-", len2) + "+" +
-              printLine("-", len3) + "+" + printLine("-", len4) + "+";
-          String line2 = "|" + printLine(" ", len1) + "+" + printLine("-", len2) + "+" +
-              printLine(" ", len3) + "+" + printLine("-", len4) + "+";
-          // table header
-          tableInfo.append(line).append("\n");
-          tableInfo.append("|" + printLine(" ", (len1 - "Module".length())) + "Module" + "|" +
-              printLine(" ", (len2 - "Operation Step".length())) + "Operation Step" + "|" +
-              printLine(" ", (len3 + len4 + 1 - "Query Cost".length())) +
-              "Query Cost" + "|" + "\n");
-          // driver part
-          tableInfo.append(line).append("\n");
-          tableInfo.append("|" + printLine(" ", len1) + "|" +
-              printLine(" ", (len2 - "SQL parse".length())) + "SQL parse" + "|" +
-              printLine(" ", len3) + "|" +
-              printLine(" ", (len4 - sql_parse_time.length())) + sql_parse_time + "|" + "\n");
-          tableInfo.append(line2).append("\n");
-          tableInfo.append("|" +printLine(" ", (len1 - "Driver".length())) + "Driver" + "|" +
-              printLine(" ", (len2 - "Load meta data".length())) + "Load meta data" + "|" +
-              printLine(" ", (len3 - driver_part_time.length())) + driver_part_time + "|" +
-              printLine(" ", (len4 - load_meta_time.length())) +
-              load_meta_time + "|" + "\n");
-          tableInfo.append(line2).append("\n");
-          tableInfo.append("|" +
-              printLine(" ", (len1 - "Part".length())) + "Part" + "|" +
-              printLine(" ", (len2 - "Block allocation".length())) +
-              "Block allocation" + "|" +
-              printLine(" ", len3) + "|" +
-              printLine(" ", (len4 - block_allocation_time.length())) +
-              block_allocation_time + "|" + "\n");
-          tableInfo.append(line2).append("\n");
-          tableInfo.append("|" +
-              printLine(" ", len1) + "|" +
-              printLine(" ", (len2 - "Block identification".length())) +
-              "Block identification" + "|" +
-              printLine(" ", len3) + "|" +
-              printLine(" ", (len4 - block_identification_time.length())) +
-              block_identification_time + "|" + "\n");
-          tableInfo.append(line).append("\n");
-
-          // once the statistics be printed, remove it from the map
-          queryStatisticsMap.remove(key);
-          // show query statistic as "query id" + "table"
-          return "Print query statistic for query id: " + key + "\n" + tableInfo.toString();
-        }
-      } catch (Exception ex) {
-        return "Put statistics into table failed, catch exception: " + ex.getMessage();
-      }
-    }
-    return null;
-  }
-
-  /**
-   * Below method will create string like "***********"
-   *
-   * @param a
-   * @param num
-   */
-  public static String printLine(String a, int num)
-  {
-    StringBuilder builder = new StringBuilder();
-    for (int i = 0; i < num; i++) {
-      builder.append(a);
-    }
-    return builder.toString();
-  }
 }
