+++ core/src/main/java/org/apache/carbondata/scan/result/iterator/DetailQueryResultIterator.java
@@ -25,8 +25,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
-import org.apache.carbondata.core.carbon.querystatistics.QueryStatistic;
-import org.apache.carbondata.core.carbon.querystatistics.QueryStatisticsConstants;
 import org.apache.carbondata.scan.executor.exception.QueryExecutionException;
 import org.apache.carbondata.scan.executor.infos.BlockExecutionInfo;
 import org.apache.carbondata.scan.model.QueryModel;
@@ -47,28 +45,11 @@ public class DetailQueryResultIterator extends AbstractDetailQueryResultIterator
 
   public DetailQueryResultIterator(List<BlockExecutionInfo> infos, QueryModel queryModel) {
     super(infos, queryModel);
-    this.queryModel = queryModel;
-  }
-
-  private Boolean flag;
-
-  private long total = 0L;
-
-  private QueryModel queryModel;
-
-  @Override public boolean hasNext() {
-    flag = super.hasNext();
-    if(!flag && total > 0) {
-      QueryStatistic statistic = new QueryStatistic();
-      statistic.addFixedTimeStatistic(QueryStatisticsConstants.SCAN_BLOCKS_TIME, total);
-      queryModel.getStatisticsRecorder().recordStatistics(statistic);
-    }
-    return flag;
   }
 
   @Override public BatchResult next() {
     BatchResult result;
-    Long startTime = System.currentTimeMillis();
+    long startTime = System.currentTimeMillis();
     try {
       if (future == null) {
         future = execute();
